EXSWHTEC-224 - Cleanup of doxygen comments

+++ catch/unit/graph/hipLaunchHostFunc.cc
@@ -27,7 +27,7 @@ THE SOFTWARE.
  * @{
  * @ingroup GraphTest
  * `hipLaunchHostFunc(hipStream_t stream, hipHostFn_t fn, void *userData)` -
- * Enqueues a host function call in a stream.
+ * enqueues a host function call in a stream
  */
 
 static void hostNodeCallbackDummy(void* data) { REQUIRE(data == nullptr); }
@@ -43,20 +43,16 @@ static void hostNodeCallback(void* data) {
 /**
  * Test Description
  * ------------------------
- *  - Validates handling of invalid arguments:
- *    -# When stream is legacy/`nullptr` stream
- *      - Platform specific (NVIDIA)
- *      - Expected output: return `hipErrorStreamCaptureImplicit`
- *    -# When function is `nullptr`
- *      - Expected output: return `hipErrorInvalidValue`
- *    -# When stream is uninitialized
- *      - Expected output: return `hipErrorContextIsDestroyed`
+ *    - Test to verify API behavior with invalid arguments:
+ *        -# Stream is legacy/nullptr stream
+ *        -# Function is nullptr
+ *        -# Stream is uninitialized
  * Test source
  * ------------------------
- *  - catch\unit\graph\hipLaunchHostFunc.cc
+ *    - catch\unit\graph\hipLaunchHostFunc.cc
  * Test requirements
  * ------------------------
- *  - HIP_VERSION >= 5.3
+ *    - HIP_VERSION >= 5.3
  */
 TEST_CASE("Unit_hipLaunchHostFunc_Negative_Parameters") {
   StreamGuard stream_guard(Streams::created);
@@ -86,14 +82,14 @@ TEST_CASE("Unit_hipLaunchHostFunc_Negative_Parameters") {
 /**
  * Test Description
  * ------------------------
- *  - Test to verify enquing a host function into a stream.
- *  - Checks if the captured computation result is correct.
+ *    - Test to verify enquing a host function into a stream, which checks if
+ * the captured computation result is correct
  * Test source
  * ------------------------
- *  - catch\unit\graph\hipLaunchHostFunc.cc
+ *    - catch\unit\graph\hipLaunchHostFunc.cc
  * Test requirements
  * ------------------------
- *  - HIP_VERSION >= 5.3
+ *    - HIP_VERSION >= 5.3
  */
 TEST_CASE("Unit_hipLaunchHostFunc_Positive_Functional") {
   LinearAllocGuard<float> A_h(LinearAllocs::malloc, sizeof(float));
@@ -121,7 +117,7 @@ TEST_CASE("Unit_hipLaunchHostFunc_Positive_Functional") {
   HIP_CHECK(hipGraphInstantiate(&graphExec, graph, nullptr, nullptr, 0));
 
   // Replay the recorded sequence multiple times
-  for (int i = 0; i < kLaunchIters; i++) {
+  for (size_t i = 0; i < kLaunchIters; i++) {
     std::fill_n(A_h.host_ptr(), 1, static_cast<float>(i));
     HIP_CHECK(hipGraphLaunch(graphExec, stream));
     HIP_CHECK(hipStreamSynchronize(stream));
@@ -139,14 +135,14 @@ static void thread_func_pos(hipStream_t* stream, hipHostFn_t fn, float** data){
 /**
  * Test Description
  * ------------------------
- *  - Test to verify enquing a host function into a stream on a different thread.
- *  - Checks if the captured computation result is correct.
+ *    - Test to verify enquing a host function into a stream on a different
+ * thread, which checks if the captured computation result is correct
  * Test source
  * ------------------------
- *  - catch\unit\graph\hipLaunchHostFunc.cc
+ *    - catch\unit\graph\hipLaunchHostFunc.cc
  * Test requirements
  * ------------------------
- *  - HIP_VERSION >= 5.3
+ *    - HIP_VERSION >= 5.3
  */
 TEST_CASE("Unit_hipLaunchHostFunc_Positive_Thread") {
   LinearAllocGuard<float> A_h(LinearAllocs::malloc, sizeof(float));
@@ -175,7 +171,7 @@ TEST_CASE("Unit_hipLaunchHostFunc_Positive_Thread") {
   HIP_CHECK(hipGraphInstantiate(&graphExec, graph, nullptr, nullptr, 0));
 
   // Replay the recorded sequence multiple times
-  for (int i = 0; i < kLaunchIters; i++) {
+  for (size_t i = 0; i < kLaunchIters; i++) {
     std::fill_n(A_h.host_ptr(), 1, static_cast<float>(i));
     HIP_CHECK(hipGraphLaunch(graphExec, stream));
     HIP_CHECK(hipStreamSynchronize(stream));
